Save the llm generated yaml playbook if available
